+++ src/ts/modules/autoPitchAccent.ts
@@ -39,6 +39,8 @@ const COLOR_TAGS = {
 
   起伏: 'kifuku',
   kifuku: 'kifuku',
+
+  'pa-none': 'none',
 } as const;
 
 // searches for first bolded
@@ -674,7 +676,7 @@ export class AutoPitchAccent extends RunnableModule {
     return str;
   }
 
-  getPAColorTag(tags: readonly string[]): PAGroup | null {
+  getPAColorTag(tags: readonly string[]): PAGroup | null | "none" {
     for (const ct of Object.keys(COLOR_TAGS)) {
       if (tags.includes(ct)) {
         return COLOR_TAGS[ct as keyof typeof COLOR_TAGS];
@@ -752,7 +754,9 @@ export class AutoPitchAccent extends RunnableModule {
     if (this.getOption('autoPitchAccent.coloredPitchAccent.enabled')) {
       if (this.attemptGlobalColor) {
         const paColorTag = this.getPAColorTag(tags);
-        if (paColorTag !== null) {
+        if (paColorTag === "none") {
+          // DO NOTHING
+        } else if (paColorTag !== null) {
           this.paintDisplay(paColorTag);
         } else if (dispPosData.mainPAGroup !== null) {
           this.paintDisplay(dispPosData.mainPAGroup);